robot_motion/robots/lerobot/requirements.txt
numpy>=2.0.0; python_version >= "3.10"
scipy>=1.10.0
PyYAML>=6.0.0
matplotlib>=3.8.0; python_version >= "3.10"
pyserial>=3.5
lcm>=1.5.0
tqdm>=4.60.0
feetech-servo-sdk>=1.0.0
draccus>=0.8.0
mujoco>=3.0.0
fsspec>=2023.0.0
spatialmath-python
pin>=3.0.0
pytest==7.0.0 
ipykernel==6.30.0